Measure timestamp width using ncurses
This allows for non-ASCII characters in timestamps, and simplifies
things by including the trailing space in the width.

+++ chat.h
@@ -283,7 +283,7 @@ enum { TimeCap = 64 };
 extern struct Time {
 	bool enable;
 	const char *format;
-	size_t width;
+	int width;
 } uiTime;
 extern struct Util uiNotifyUtil;
 void uiInitEarly(void);
